assignment_flask2/requirements.txt
flask
pymongo
python-dotenv
dnspython
json